+++ client/blocks/payment-request/custom-button.js
@@ -6,7 +6,7 @@ import { __ } from '@wordpress/i18n';
 /**
  * Internal dependencies
  */
-import { getStripeServerData } from '../stripe-utils';
+import { getStripeServerData } from '../../stripe-utils';
 
 export const CustomButton = ( { onButtonClicked } ) => {
 	const {
