+++ client/blocks/payment-request/payment-request-express.js
@@ -10,7 +10,7 @@ import {
 /**
  * Internal dependencies
  */
-import { getStripeServerData } from '../stripe-utils';
+import { getStripeServerData } from '../../stripe-utils';
 import { GooglePayButton, shouldUseGooglePayBrand } from './branded-buttons';
 import { CustomButton } from './custom-button';
 import {
