+++ client/stripe-utils/load-stripe.js
