+++ client/stripe-utils/normalize.js
